+++ code/lobbyscraper.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/bin/env python2
 # -*- coding: utf-8 -*-
 
 """
@@ -7,158 +7,157 @@ and converts the relevant data into a json file.
 """
 
 import re
-from datetime import datetime, date, time
+from datetime import datetime, date
+import time
 import json
 import os
 from bs4 import BeautifulSoup
 import urllib2
-import pandas as pd
+import dataset
 
 __author__ = "Stefan Kasberger"
 __copyright__ = "Copyright 2015"
 __license__ = "MIT"
-__version__ = "0.0.1"
+__version__ = "0.2"
 __maintainer__ = "Stefan Kasberger"
 __email__ = "mail@stefankasberger.at"
-__status__ = "Development" # 'Development', 'Production' or 'Prototype'
+__status__ = "Production" # 'Development', 'Production' or 'Prototype'
 
 ###    GLOBAL   ###
 
-BASE_URL = 'http://www.lobbyreg.justiz.gv.at/edikte/ir/iredi18.nsf'
-QUERY_URL = BASE_URL+'/liste!OpenForm&subf=a'
 ROOT_FOLDER = os.path.dirname(os.getcwd())
-FOLDER_HTML = ROOT_FOLDER + '/data/raw/'
+FOLDER_RAW_HTML = ROOT_FOLDER + '/data/raw/html/'
+FOLDER_RAW_PDF = ROOT_FOLDER + '/data/raw/pdf/'
 FOLDER_JSON = ROOT_FOLDER + '/data/json/'
 FOLDER_CSV = ROOT_FOLDER + '/data/csv/'
-FILENAME_HTML = 'lobbyingregister.htm'
-FILENAME_JSON = 'lobbyingregister.json'
-FILENAME_CSV = 'lobbyingregister.csv'
+FILENAME_BASE = 'lobbyingregister'
+BASE_URL = 'http://www.lobbyreg.justiz.gv.at/edikte/ir/iredi18.nsf'
+BASE_URL_ATTACHMENTS = 'http://www.lobbyreg.justiz.gv.at'
+QUERY_URL = BASE_URL+'/liste!OpenForm&subf=a'
+DELAY_TIME = 2 # in seconds
+# TS = datetime.now().strftime('%Y-%m-%d-%H-%M')
+TS = '2016-04-19-19-25'
 
 ###    FUNCTIONS   ###
 
 def SetupEnvironment():
-    if not os.path.exists(FOLDER_HTML):
-        os.makedirs(FOLDER_HTML)
+    """Sets up the folder structure and working environment.
+    """
+    if not os.path.exists(FOLDER_RAW_HTML):
+        os.makedirs(FOLDER_RAW_HTML)
+    if not os.path.exists(FOLDER_RAW_PDF):
+        os.makedirs(FOLDER_RAW_PDF)
     if not os.path.exists(FOLDER_JSON):
         os.makedirs(FOLDER_JSON)
     if not os.path.exists(FOLDER_CSV):
         os.makedirs(FOLDER_CSV)
 
 def FetchHtml(url):
-    """Fetches html url via urllib().
+    """Fetches html url via urllib2.
     
     Args:
-        url: url to fetch
+        url: url to fetch (string).
     
     Returns:
-        html string as unicode
+        html: html string as unicode
     """
     response = urllib2.urlopen(url)
-    rawHtml = response.read().decode('utf-8')
-    return rawHtml
+    html = response.read().decode('utf-8')
+    time.sleep(DELAY_TIME)
+    
+    return html
 
-def FetchHtmlList(url, folder, filename):
+def FetchHtmlOverview(url, folder):
     """Fetches html from the overview list of the lobbyingregister entries and saves it locally.
     
     Args:
-        url: url to fetch
-        folder: to save the html
-        filename: filename for the html file
-    
-    Returns:
-        html string
+        url: url to fetch (string).
+        folder: directory to save the html files in.    
     """
     rawHtml = FetchHtml(url)
+
     if not os.path.exists(folder):
         os.makedirs(folder)
-    Save2File(rawHtml, folder+'/'+filename)
-    return rawHtml
+    Save2File(rawHtml, folder+'/overview-page.htm')
 
-def FetchHtmlOrganisations(organisations, folder):
-    """Fetches html from a lobbying-organisation and saves it locally.
+def FetchHtmlEntries(entries, folder):
+    """Fetches html from an entry in the table and saves it locally with the unique id as postfix.
     
     Args:
-        organisations: dict with sequencial id's of organisations as keys.
-        folder: to save the html
-    
-    Returns:
-        dict() of sequencial id's of organisations as key and html as value.
+        entries: list with sequencial dict() of organisations.
+        folder: to save the html in.
     """
-    html = {}
-    for id in organisations.keys():
-        html[id] = FetchHtml(organisations[id]['url'])
-        Save2File(html[id], folder+'/'+str(id)+'.htm')
-    return html
+    for entry in entries:
+        html = FetchHtml(entry['url'])
+        Save2File(html, folder+'/entry-'+str(entry['ID'])+'.htm')
 
 def Save2File(data, filename):
-    """Saves file locally
+    """Saves data on specified place on harddrive.
     
     Args:
-        data: string to save
-        filename: name of the file
-    
-    Returns:
-        na
+        data: string to save.
+        filename: string of the filepath.
     """
     text_file = open(filename, "w")
     text_file.write(data.encode('utf-8'))
     text_file.close()
 
-def Save2CSV(data, filename):
-    """Exports the dict into a csv file. Uses pandas dataframe and write csv functions.
-    
-    Args:
-        data: dict() with all the data
-        filename: name of the file with folder
-    
-    Returns:
-        na
-    """
-    data = pd.DataFrame(data)
-    data = data.T
-    data.to_csv(filename, sep=';', encoding='utf-8')
-
 def ReadFile(filename):
-    """Reads file and returns the html.
+    """Reads file and returns the text.
     
     Args:
         filename: name of the file
     
     Returns:
-        html from the file.
+        string: content of file as string
     """
     f = open(filename, 'r')
-    html = f.read()
-    return html
+    string = f.read()
 
-def ReadOrganisations(folder):
-    """Reads in all html-files from the organisations folder.
+    return string
+
+def ReadEntryFilesInFolder(folder):
+    """Reads-in all entry html-files from specified folder.
     
     Args:
-        folder: folder where the organisation html-files are stored.
+        folder: folder where the html-files are stored in.
     
     Returns:
-        dict() of sequencial id's of organisations as key and html as value.
+        sortedList: list[] of html texts sorted by file-postfix, which is the unique id of the entry from the table.
     """
-    html = {}
+    htmlList = []
+    sortedList = []
+
     for filename in os.listdir(folder):
-        rawHtml = ReadFile(folder+'/'+filename)
-        if not filename in FILENAME_HTML:
-            html[int(filename.split('.')[0])] = rawHtml
-    return html
+        if filename.find('entry-') >= 0:
+            rawHtml = ReadFile(folder+'/'+filename)
+            fileIndex = filename.split('.')[0].split('-')[1]
+            htmlList.append((int(fileIndex), rawHtml))
+    
+    # sort list of duppels after first element (the filename postfix) and store to list[]
+    htmlList = sorted(htmlList, key=lambda htmlList: htmlList[0])
+    for idx, html in htmlList:
+        sortedList.append(html)
 
-def ParseList(html, timestamp):
-    """Parses the needed facts out of the overview list html.
+    return sortedList
+
+def ParseTable(html):
+    """Parses the needed facts out of the overview table.
     
     Args:
-        html: html string 
-        timestamp: time when the html download was started.
+        html: html (string) to parse. 
     
     Returns:
-        dict() of sequencial id's of organisations as key and dict() with facts as value.
+        list[] of dict() of entries.
+            'ID': serial number created from scraper when parsing through the table rows.
+            'entryDescription': description of the entry.
+            'lobbyingOrgaType': type of lobbying organisation (A1, A2, B, C, D).
+            'url': url of the detail page.
+            'lastUpdate': last update of detail page.
+            'registryNumber': number of company in lobbying-register.
     """
-    lobbyList = {}
-    counter = 0
+    lobbyList = []
+    counter = 1
 
     soup = BeautifulSoup(html)
     
@@ -167,158 +166,361 @@ def ParseList(html, timestamp):
         tds = tr.find_all('td')
 
         # assign variables from html table to dict
-        organisation = {}
-        organisation['description'] =  unicode(tds[1].string) # organisation
-        organisation['registry-department'] =  tds[3].string # register department
-        organisation['url'] =  BASE_URL+'/'+tds[2].a['href'] # register number url
-        organisation['last-update'] = str(datetime.strptime(tds[5].string, '%d.%m.%Y')) # last update
-        organisation['register-number'] = tds[2].string
+        entry = {}
+        entry['ID'] =  str(counter)
+        entry['entryDescription'] =  unicode(tds[1].string)
+        entry['lobbyingOrgaType'] =  tds[3].string
+        entry['url'] =  BASE_URL+'/'+tds[2].a['href']
+        entry['lastUpdate'] = str(datetime.strptime(tds[5].string, '%d.%m.%Y'))
+        entry['registryNumber'] = tds[2].string
         
-        lobbyList[counter] = organisation
+        lobbyList.append(entry)
         counter += 1
+
     return lobbyList
 
-def ParseOrganisations(htmlList, organisations):
+def ParseEntries(htmlList, entries):
     """Parses the needed facts out of the organisation html.
     
     Args:
         htmlList: list() of html strings.
-        organisations: dict() of sequencial id's of organisations as key and dict() with facts as value.
+        entries: list() of dict() of entries.
     
     Returns:
-        dict() of sequencial id's of organisations as key and dict() with facts as value.
+        list[] of dict() of entries.
+            'ID': serial number created from scraper when parsing through the table rows.
+            'entryDescription': description of the entry.
+            'lobbyingOrgaType': type of lobbying organisation (A1, A2, B, C, D).
+            'url': url of the detail page.
+            'lastUpdate': last update of detail page.
+            'registryNumber': number of company in lobbying-register.
+            'orgaName': name of organisation.
+            'companyRegisterNumber': number of the national company register.
+            'businessActivities': area in which the organisation is active in business.
+            'postalAddress': postal address.
+            'registeredOffice': Place, where the company is officially registered.
+            'businessYearStart': Day, when the business year starts.
+            'legalFoundation': 
+            'codeOfConduct': Code of conduct.
+            'website': url of the website
+            'lobbyists': list of lobbyists.
+            'lobbyingRevenue': Revenue from lobbying.
+            'lobbyingRequests': Number of lobbying requests.
+            'numLobbyists': Number of lobbyists.
+            'lobbyingCostsGreater100000': 
+            'lobbyingCosts': Costs of lobbying.
+            'suborganisations': List of suborganisations.
+            'attachmentUrls': url to an attachment.
+            'comment': comment to precise other fields.
     """
+    entriesList = []
 
-    for id in organisations.keys():
-        soup = BeautifulSoup(htmlList[id])
+    for entry in entries:
+        soup = BeautifulSoup(htmlList[int(entry['ID'])-1])
         html = unicode(soup)
        
-        # regex type of registry department: B, C 
+        # regex testing type of registry department: B, C 
         regDepartment = re.findall(r'Registerabteilung:</strong></dt>\n<dd><strong>(.*)</strong></dd></dl>', html)
         if regDepartment:
-            if organisations[id]['registry-department'] != regDepartment[0]:
-                print 'ERROR: register department differs!'
+            if entry['lobbyingOrgaType'] != regDepartment[0]:
+                print 'ERROR: lobbying organisation type differs!'
 
-        # regex register number: B, C
+        # regex testing register number: B, C
         regNum = re.findall(r'Registerzahl:</strong></dt>\n<dd><strong>(.*)</strong></dd></dl>', html)
         if regNum:
-            if organisations[id]['register-number'] != regNum[0]:
-                print 'ERROR: register number differs!'
+            if entry['registryNumber'] != regNum[0]:
+                print 'ERROR: company register number differs!'
 
         # regex name: A1, B, C
         name = re.findall(r'Name.*:</strong></dt>\n<dd><strong>(.*)</strong></dd></dl>', html)
         if name:
-            organisations[id]['name'] = name[0]
+            entry['orgaName'] = name[0]
 
         # regex date announced: A1, C
         announceDate = re.findall(r'Bekannt gemacht am:</dt>\n<dd>(.*)</dd>', html)
         if announceDate:
-            organisations[id]['date-announced'] = str(datetime.strptime(announceDate[0], '%d.%m.%Y'))
+            entry['dateAnnounced'] = str(datetime.strptime(announceDate[0], '%d.%m.%Y'))
 
-        # regex last update: A1, B, C
+        # regex testing last update: A1, B, C
         lastUpdate = re.findall(r'Letzte .*nderung:</dt>\n<dd>(.*)</dd>', html)
         if lastUpdate:
-            if organisations[id]['last-update'] != str(datetime.strptime(lastUpdate[0], '%d.%m.%Y')):
+            if entry['lastUpdate'] != str(datetime.strptime(lastUpdate[0], '%d.%m.%Y')):
                 print "ERROR: register last update differs!"
 
-        # regex corporate-number: A1
+        # regex corporate-number: A1, B
         corporateNumber = re.findall(r'Firmenbuchnummer:</dt>\n<dd>(.*)</dd>', html)
         if corporateNumber:
-            organisations[id]['corporate-number'] = corporateNumber[0]
+            entry['companyRegisterNumber'] = corporateNumber[0]
 
         # regex registered office address: A1, C, D
         regOfficeAddress = re.findall(r'itz:</dt>\n<dd>(.*)</dd></dl>', html)
         if regOfficeAddress:
-            organisations[id]['registered-office-address'] = regOfficeAddress[0]
+            entry['registeredOffice'] = regOfficeAddress[0]
 
         # regex mail address: A1, C, D
-        mailAddress = re.findall(r'nschrift:</dt>\n<dd>(.*)</dd></dl>', html)
-        if mailAddress:
-            organisations[id]['mailing-address'] = mailAddress[0]
+        postalAddress = re.findall(r'nschrift:</dt>\n<dd>(.*)</dd></dl>', html)
+        if postalAddress:
+            entry['postalAddress'] = postalAddress[0]
 
-        # regex start business year: A1
+        # regex start business year: A1, B
         startBusinessYear = re.findall(r'ftsjahres:</dt>\n<dd>(.*)</dd></dl>', html)
         if startBusinessYear:
-            organisations[id]['start-business-year'] = str(datetime.strptime(startBusinessYear[0], '%d.%m.'))
+            entry['businessYearStart'] = startBusinessYear[0]
 
         # regex legal foundation: C
         legalFoundation = re.findall(r'Gesetzliche Grundlage:</dt>\n<dd>(.*)</dd></dl>', html)
         if legalFoundation:
-            organisations[id]['legal-foundation'] = legalFoundation[0]
+            entry['legalFoundation'] = legalFoundation[0]
 
-        # regex area of activities: A1, B, C, D
+        # regex area of activities: A1, B, D
         areaActivities = re.findall(r'bereich:</dt>\n<dd>(.*)</dd></dl>', html)
         if areaActivities:
-            organisations[id]['activities'] = areaActivities[0]
+            entry['businessActivities'] = areaActivities[0]
 
-        # regex codex: A1
+        # regex codex: A1, B
         codex = re.findall(r'Verhaltenskodex:</dt>\n<dd>(.*)</dd></dl>', html)
         if codex:
-            organisations[id]['codex'] = codex[0]
+            entry['codeOfConduct'] = codex[0]
 
         # regex website: A1, B, C, D
         website = re.findall(r'Homepage:</dt>\n<dd><a href="(.*)" target="_blank">.*</a></dd></dl>', html)
         if website:
-            organisations[id]['website'] = website[0]
+            entry['website'] = website[0]
 
         # regex lobbyists: A1
-        lobbyists = re.findall(r'obbyist:</dt>\n<dd>(.*)</dd></dl>', html)
+        lobbyists = re.findall(r'obbyist.*:</dt>\n<dd>(.*)</dd></dl>', html)
         if lobbyists:
-            organisations[id]['lobbyists'] = lobbyists[0].split('<br/>')
+            entry['lobbyists'] = lobbyists[0].split('<br/>')
 
-        # regex lobbying revenue: A1
+        # regex lobbying revenue: A1, B
         lobbyingRevenue = re.findall(r'Lobbying-Umsatz:</dt>\n<dd>(.*)</dd></dl>', html)
         if lobbyingRevenue:
-            organisations[id]['lobbying-revenue'] = lobbyingRevenue[0]
+            entry['lobbyingRevenue'] = lobbyingRevenue[0]
 
         # regex lobbying request: A1
         lobbyingRequest = re.findall(r'<dt title="Anzahl der bearbeiteten Lobbying-Auftr.*ge:</dt>\n<dd>(.*)</dd></dl>', html)
         if lobbyingRequest:
-            organisations[id]['lobbying-request'] = lobbyingRequest[0]
+            entry['lobbyingRequests'] = lobbyingRequest[0]
 
         # regex number of lobbyists: C, D
         numLobbyists = re.findall(r'Anzahl Interessenvertreter:</dt>\n<dd>(.*)</dd>', html)
         if numLobbyists:
-            organisations[id]['number-lobbyists'] = numLobbyists[0]
+            entry['numLobbyists'] = numLobbyists[0]
 
-        # regex costs lobbying: B, C
+        # regex costs lobbying: B
         costsB = re.findall(r'Lobbying-Aufwand > EUR 100.000:</dt>\n<dd>(.*)</dd>', html)
-        costsC = re.findall(r'Kosten der Interessenvertretung:</dt>\n<dd>(.*)</dd>', html)
         if costsB:
             if costsB[0] == 'Ja':
-                organisations[id]['costs-lobbying-greater-100000'] = True
+                entry['lobbyingCostsGreater100000'] = True
             if costsB[0] == 'Nein':
-                organisations[id]['costs-lobbying-greater-100000'] = False
+                entry['lobbyingCostsGreater100000'] = False
+        # regex costs lobbying: C, D
+        costsC = re.findall(r'Kosten der Interessenvertretung:</dt>\n<dd>(.*)</dd>', html)
         if costsC:
-            organisations[id]['costs-lobbying'] = costsC[0]
+            entry['lobbyingCosts'] = costsC[0]
 
         # regex atttachments: C
-        attachment = re.findall(r'nge:</dt>\n<dd><a href="(.*)" .* target=_new>.*</a></dd>', html)
-        if attachment:
-            organisations[id]['pdf-attachment'] = BASE_URL + attachment[0]
+        attDiv = soup.find(id='UplDL')
+        if attDiv:
+            entry['attachmentUrlss'] = []
+            for link in attDiv.find_all('a'):
+                entry['attachmentUrlss'].append(BASE_URL_ATTACHMENTS + link.get('href'))
 
-        # regex suborganisations: C
+        # regex subentries: C, D
         subOrganisations = re.findall(r'<dt>Unterorganisation.*</dt>\n<dd>(.*)</dd></dl>', html)
         if subOrganisations:
-            organisations[id]['sub-organisations'] = subOrganisations[0]
+            entry['suborganisations'] = subOrganisations[0]
 
-    return organisations
+        # regex comment: C, D
+        comment = re.findall(r'<dt>Kommentar:</dt>\n<dd>(.*)</dd></dl>', html)
+        if comment:
+            entry['comment'] = comment[0]
+
+        entriesList.append(entry)
+
+    return entriesList
+
+def Save2CSV(entries, filename):
+    """Exports the dict into a csv file.
+    
+    Args:
+        entries: list of dict() with all the lobbying register entries.
+        filename: name of the file with folder.
+    """
+    csvString = '"ID","entryDescription","orgaName","businessActivities","lobbyingOrgaType","lobbyists","lobbyingRevenue","lobbyingRequests","numLobbyists","lobbyingCostsGreater100000","lobbyingCosts","registryNumber","companyRegisterNumber","suborganisations","legalFoundation","codeOfConduct","registeredOffice","website","postalAddress","lastUpdate","dateAnnounced","businessYearStart","url","attachmentUrls","comment"\n'
+
+    # iterate over each entry
+    for entry in entries:
+        uniqueId = '""'
+        entryDescription = '""'
+        orgaName = '""'
+        busActivities = '""'
+        lobOrgaType = '""'
+        lobbyists = '""'
+        lobbyingRevenue = '""'
+        lobbyingRequest = '""'
+        numLobbyists = '""'
+        costsB = '""'
+        costsC = '""'
+        regNum = '""'
+        compRegNumber = '""'
+        subOrganisations = '""'
+        legalFoundation = '""'
+        codex = '""'
+        regOfficeAddress = '""'
+        lastUpdate = '""'
+        website = '""'
+        postalAddress = '""'
+        lastUpdate = '""'
+        announceDate = '""'
+        startBusinessYear = '""'
+        url = '""'
+        attachments = '""'
+        comment = '""'
+
+        # read out each attribute
+        for elem in entry.keys():
+            val = entry[elem]
+            if elem == 'ID':
+                uniqueId = '"'+val+'"'
+            if elem == 'entryDescription':
+                entryDescription = '"'+val+'"'
+            if elem == 'orgaName':
+                orgaName = '"'+val+'"'
+            if elem == 'businessActivities':
+                busActivities = '"'+val.replace('"', '').replace('\n', '').replace('\r', '')+'"'
+            if elem == 'lobbyingOrgaType':
+                lobOrgaType = '"'+val+'"'
+            if elem == 'lobbyists':
+                lobbyists = '"'
+                for lob in val:
+                    lobbyists += lob+', '
+                lobbyists = lobbyists[:-1]
+                lobbyists += '"'
+            if elem == 'lobbyingRevenue':
+                lobbyingRevenue = '"'+val+'"'
+            if elem == 'lobbyingRequests':
+                lobbyingRequest = '"'+val+'"'
+            if elem == 'numLobbyists':
+                numLobbyists = '"'+val+'"'
+            if elem == 'lobbyingCostsGreater100000':
+                costsB = '"'+val+'"'
+            if elem == 'lobbyingCosts':
+                costsC = '"'+val+'"'
+            if elem == 'registryNumber':
+                regNum = '"'+val+'"'
+            if elem == 'companyRegisterNumber':
+                compRegNumber = '"'+val+'"'
+            if elem == 'suborganisations':
+                subOrganisations = '"'+val+'"'
+            if elem == 'legalFoundation':
+                legalFoundation = '"'+val+'"'
+            if elem == 'codeOfConduct':
+                codex = '"'+val.replace('"', '').replace('\n', '').replace('\r', '')+'"'
+            if elem == 'registeredOffice':
+                regOfficeAddress = '"'+val+'"'
+            if elem == 'website':
+                website = '"'+val+'"'
+            if elem == 'postalAddress':
+                postalAddress = '"'+val+'"'
+            if elem == 'lastUpdate':
+                lastUpdate = '"'+val+'"'
+            if elem == 'dateAnnounced':
+                announceDate = '"'+val+'"'
+            if elem == 'businessYearStart':
+                startBusinessYear = '"'+val+'"'
+            if elem == 'url':
+                url = '"'+val+'"'
+            if elem == 'attachmentUrlss':
+                attachments = '"'
+                for att in val:
+                    attachments += att+', '
+                attachments +='"'
+            if elem == 'comment':
+                comment = '"'+val+'"'
+
+        csvString += uniqueId+','+entryDescription+','+orgaName+','+busActivities+','+lobOrgaType+','+lobbyists+','+lobbyingRevenue+','+lobbyingRequest+','+numLobbyists+','+costsB+','+costsC+','+regNum+','+compRegNumber+','+subOrganisations+','+legalFoundation+','+codex+','+regOfficeAddress+','+website+','+postalAddress+','+lastUpdate+','+announceDate+','+startBusinessYear+','+url+','+attachments+','+comment+'\n'
+
+    Save2File(csvString, filename)
+    print 'Lobbying data exported as CSV:',filename
+
+def FetchAttachments(entries, folder):
+    """Fetches all attachments from the lobbying-register entries.
+    
+    Args:
+        entries: list[] of dict() with all the lobbying-register data.
+        folder: directory, where the files are stored in.
+    """
+    for entry in entries:
+        if 'attachmentUrls' in entry.keys():
+            for url in entry['attachmentUrls']:
+                DownloadFile(url, folder+'/attachment-'+entry['ID']+'_'+url.split('/')[-1])
+
+def DownloadFile(url, filename):
+    """Downloads and stores an attachment.
+    
+    Args:
+        url: url of the attachment (string).
+        filename: full filepath for the saving.
+    """
+    if not os.path.exists(os.path.dirname(os.path.abspath(filename))):
+        os.makedirs(os.path.dirname(os.path.abspath(filename)))
+    response = urllib2.urlopen(url.encode('utf-8'))
+    file = open(filename, 'w')
+    file.write(response.read())
+    file.close()
+    time.sleep(DELAY_TIME)
+
+def Save2SQLite(lobbyEntries):
+    """Saves the lobbing register entries in a SQLite database. This is not working, because of key-value issues of the dicts().
+    
+    Args:
+        lobbyEntries: list[] of dicts() with lobbying register entries.
+    """
+    db = dataset.connect('sqlite:///:memory:')
+    table = db['lobbyRegisterAT']
+    for entry in lobbyEntries:
+        print entry
+        table.insert(entry)
+
+    # Wien = table.find_one(registeredOffice='Wien')
+    # print Wien
 
 ###    MAIN   ###
 
 if __name__ == '__main__':
+    
+    # setup
+    startTime = datetime.now()
+    print 'start:', startTime
     SetupEnvironment()
-    # ts = datetime.now().strftime('%Y-%m-%d-%H-%M')
-    ts = '2015-05-05-00-37'
-    print ts
-    # htmlList = FetchHtmlList(QUERY_URL, ROOT_FOLDER+'/data/raw/'+ts, FILENAME_HTML) # list(html as text)
-    htmlList = ReadFile(FOLDER_HTML+ts+'/'+FILENAME_HTML) # list(html as text)
-    lobbyList = ParseList(htmlList, ts) # dict(registry-number: dict(url, type, description, etc))
-    Save2File(json.dumps(lobbyList, indent=2, ensure_ascii=False), FOLDER_JSON+ts+'_'+FILENAME_JSON)
-    # htmlOrgas = FetchHtmlOrganisations(lobbyList, ROOT_FOLDER+'/data/raw/'+ts) # dict(registry-number: html)
-    htmlOrgas = ReadOrganisations(FOLDER_HTML+ts) # dict(registry-number: html)
-    lobbyOrgas = ParseOrganisations(htmlOrgas, lobbyList)
-    Save2File(json.dumps(lobbyOrgas, indent=2, ensure_ascii=False), FOLDER_JSON+ts+'_'+FILENAME_JSON)
-    Save2CSV(lobbyList, FOLDER_CSV+ts+'_'+FILENAME_CSV)
-    # scraperwiki.sqlite.save(unique_keys=['', ''], data=data)
+    DOWNLOAD_FILES = False
+    PARSE_FILES = False
+    DOWNLOAD_ATTACHMENTS = False
+    EXPORT_DATA = False
+ 
+    if DOWNLOAD_FILES:
+        FetchHtmlOverview(QUERY_URL, FOLDER_RAW_HTML+TS) 
+        htmlOverview = ReadFile(FOLDER_RAW_HTML+TS+'/overview-page.htm')
+        lobbyList = ParseTable(htmlOverview) 
+        Save2File(json.dumps(lobbyList, indent=2, ensure_ascii=True), FOLDER_JSON+TS+'_'+FILENAME_BASE+'.json')
+        FetchHtmlEntries(lobbyList, FOLDER_RAW_HTML+TS) 
+ 
+    if PARSE_FILES:
+        htmlOverview = ReadFile(FOLDER_RAW_HTML+TS+'/overview-page.htm')
+        lobbyList = ParseTable(htmlOverview) 
+        htmlEntries = ReadEntryFilesInFolder(FOLDER_RAW_HTML+TS) 
+        lobbyEntries = ParseEntries(htmlEntries, lobbyList)
+        Save2File(json.dumps(lobbyEntries, indent=2, ensure_ascii=True), FOLDER_JSON+TS+'_'+FILENAME_BASE+'.json')
+    
+    if DOWNLOAD_ATTACHMENTS:    
+        lobbyEntries = ReadFile(FOLDER_JSON+TS+'_'+FILENAME_BASE+'.json')
+        FetchAttachments(lobbyEntries, FOLDER_RAW_PDF+TS)
+ 
+    if EXPORT_DATA:
+        lobbyEntries = json.loads(ReadFile(FOLDER_JSON+TS+'_'+FILENAME_BASE+'.json'))
+        Save2CSV(lobbyEntries, FOLDER_CSV+TS+'_'+FILENAME_BASE+'.csv')
+        # Save2SQLite(lobbyEntries) # does not run!
+
+    print 'runtime:', (datetime.now() - startTime)