+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### HEAD
-* 
 
-### 0.0.1: May 4th, 2015
-* init repo
+### Version 0.2: 2016-04-26
+- update documentation: commenting code, update README.md
+- add csv export
+- re-factor code
+- add attachment download
+- add SQLite export
+
+### Version 0.1: 2015-05-04
+**init repo**